+++ package.json
@@ -31,6 +31,8 @@
     "eslint-plugin-mozilla": "0.2.3",
     "eslint-plugin-react": "^6.7.1",
     "husky": "^0.12.0",
+    "react": "=15.3.2",
+    "react-dom": "=15.3.2",
     "react-hot-loader": "^1.3.1",
     "stylelint": "^7.4.2"
   },
